feat:#39 생성한 컴포넌트로 적용
